+++ xcube_zenodo/preload.py
@@ -1,12 +1,12 @@
 # The MIT License (MIT)
-# Copyright (c) 2024 by the xcube development team and contributors
+# Copyright (c) 2024-2025 by the xcube development team and contributors
 #
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
 #
 # The above copyright notice and this permission notice shall be included in all
 # copies or substantial portions of the Software.
@@ -20,298 +20,175 @@
 # SOFTWARE.
 
 import shutil
-import threading
-import time
-from typing import Callable
 import tarfile
 import zipfile
 
 import fsspec
-import numpy as np
-import tabulate
 import requests
-import xarray as xr
-from xcube.core.store import DataStoreError
-from xcube.core.store import MutableDataStore
+from xcube.core.store import DataStoreError, MutableDataStore
+from xcube.core.store.preload import ExecutorPreloadHandle, PreloadState, PreloadStatus
 
-from ._utils import identify_file_format
-from ._utils import translate_data_id2uri
-from ._utils import is_zarr_directory
-from .constants import LOG
-
-
-class Event:
-
-    def __init__(self, data_id: str, total_size: int | float):
-        self.data_id = data_id
-        self.status = "Not started"
-        self.progress = 0.0
-        self.message = "Preloading not started yet."
-        self.total_size = total_size
-        self._callback = None
-
-    def subscribe(self, callback: Callable[[], None]):
-        self._callback = callback
-
-    def notify(self):
-        if self._callback is not None:
-            self._callback()
-
-    def update(self, status: str, progress: float, message: str):
-        self.status = status
-        self.progress = progress
-        self.message = message
-        self.notify()
+from ._utils import identify_compressed_file_format, translate_data_id2uri
+from .constants import (
+    DOWNLOAD_FOLDER,
+    PRELOAD_DECOMPRESSION_FRACTION,
+    PRELOAD_DOWNLOAD_FRACTION,
+    PRELOAD_PROCESSING_FRACTION,
+)
 
 
-class PreloadHandle:
+class ZenodoPreloadHandle(ExecutorPreloadHandle):
 
     def __init__(self, cache_store: MutableDataStore, *data_ids: str, **preload_params):
-        self._is_cancelled = False
-        self._is_closed = False
         self._cache_store = cache_store
         self._cache_fs: fsspec.AbstractFileSystem = cache_store.fs
         self._cache_root = cache_store.root
         self._data_ids = data_ids
-        self._preload_params = preload_params
-        self._download_folder_name = "downloads"
+        self._io_id = f"dataset:zarr:{self._cache_store.protocol}"
         self._download_folder = self._cache_fs.sep.join(
-            [self._cache_root, self._download_folder_name]
+            [self._cache_root, DOWNLOAD_FOLDER]
         )
-        self._events = [Event(data_id, np.nan) for data_id in data_ids]
-        self.lock = threading.Lock()
-        if preload_params.get("monitor_preload"):
-            for event in self._events:
-                event.subscribe(self._monitor_preload)
-        self._thread_download = None
-        self._thread_decompress = None
-        self._thread_prepare = None
+        self._clean_up()
+        super().__init__(data_ids=data_ids, **preload_params)
 
-    @property
-    def is_cancelled(self) -> bool:
-        return self._is_cancelled
+    def close(self) -> None:
+        self._clean_up()
 
-    def cancel(self):
-        self._is_cancelled = True
-        self._thread_prepare.join()
-        self._thread_decompress.join()
-        self._thread_download.join()
-        for event in self._events:
-            event.update("Canceled", np.nan, "Preload has been canceled by user.")
-        self.close()
+    def preload_data(self, data_id: str):
+        self._download_data(data_id)
+        self._decompress_data(data_id)
+        self._prepare_data(data_id)
 
-    @property
-    def is_closed(self) -> bool:
-        return self._is_closed
-
-    def close(self):
-        self._is_closed = True
-        for event in self._events:
-            if event.status != "Preloaded":
-                if self._cache_store.has_data(event.data_id):
-                    self._cache_store.delete_data(event.data_id)
-                else:
-                    record, filename = event.data_id.split("/")
-                    format_id = identify_file_format(event.data_id)
-                    dirname = filename.replace(f".{format_id}", "")
-                    data_id_mod = f"{record}/{dirname}"
-                    list_data_ids = self._cache_store.list_data_ids()
-                    list_data_ids_mod = [
-                        data_id for data_id in list_data_ids if data_id_mod in data_id
-                    ]
-                    for data_id in list_data_ids_mod:
-                        self._cache_store.delete_data(data_id)
-        if self._cache_fs.isdir(self._download_folder):
-            shutil.rmtree(self._download_folder)
-
-    def _monitor_preload(self):
-        rows = [
-            [
-                event.data_id,
-                event.status,
-                f"{event.progress * 100:.2f}%",
-                event.message,
-            ]
-            for event in self._events
-        ]
-        if is_jupyter():
-            import IPython.display
-
-            table = tabulate.tabulate(
-                rows,
-                headers=["Data ID", "Status", "Progress", "Message"],
-                tablefmt="html",
-            )
-            IPython.display.clear_output(wait=True)
-            IPython.display.display(table)
-
-    def preload_data(self, *data_ids: str, **preload_params):
-        self._download_data(*data_ids)
-        self._decompress_data(*data_ids)
-        self._prepare_data(*data_ids, **preload_params)
-
-    def _download_data(self, *data_ids: str):
-        # get first total size of all datasets and initialize it as Events
-        for i, data_id in enumerate(data_ids):
-            uri = translate_data_id2uri(data_id)
-            with requests.get(uri, stream=True) as response:
-                if not response.ok:
-                    raise DataStoreError(response.raise_for_status())
-                self._events[i].total_size = int(
-                    response.headers.get("content-length", 0)
-                )
+    def _download_data(self, data_id: str):
+        # get first total size of the dataset
+        uri = translate_data_id2uri(data_id)
+        with requests.get(uri, stream=True) as response:
+            _check_requests_response(response)
+            total_size = int(response.headers.get("content-length", 0))
 
         # start downloading
         chunk_size = 1024 * 1024
-
-        def download():
-            for i, data_id in enumerate(data_ids):
-                download_size = 0
-                self._events[i].update("Download started", 0, "")
-                uri = translate_data_id2uri(data_id)
-                with requests.get(uri, stream=True) as response:
-                    if not response.ok:
-                        raise DataStoreError(response.raise_for_status())
-                    record, filename = data_id.split("/")
-                    record_folder = self._cache_fs.sep.join(
-                        [self._download_folder, record]
-                    )
-                    if not self._cache_fs.isdir(record_folder):
-                        self._cache_fs.makedirs(record_folder)
-                    download_path = self._cache_fs.sep.join([record_folder, filename])
-                    with open(download_path, "wb") as file:
-                        for chunk in response.iter_content(chunk_size=chunk_size):
-                            file.write(chunk)
-                            download_size += len(chunk)
-                            self._events[i].update(
-                                "Download started",
-                                download_size / self._events[i].total_size,
-                                "",
-                            )
-                            if self._is_cancelled:
-                                break
-                self._events[i].update("Downloaded", 1.0, "")
-
-        self._thread_download = threading.Thread(
-            target=download, daemon=True, name="download_data"
+        download_size = 0
+        self.notify(
+            PreloadState(
+                data_id,
+                status=PreloadStatus.started,
+                progress=0.0,
+                message="Download in progress",
+            )
         )
-        self._thread_download.start()
-
-    def _decompress_data(self, *data_ids):
-        def decompress():
-            for i, data_id in enumerate(data_ids):
-                while not self._events[i].status == "Downloaded":
-                    time.sleep(1)
-                self._events[i].update("Decompression started", np.nan, "")
-                record, filename = data_id.split("/")
-                file_path = self._cache_fs.sep.join(
-                    [self._download_folder, record, filename]
-                )
-                if zipfile.is_zipfile(file_path):
-                    with zipfile.ZipFile(file_path, "r") as zip_ref:
-                        dirname = filename.replace(".zip", "")
-                        extract_dir = self._cache_fs.sep.join(
-                            [self._download_folder, record, dirname]
-                        )
-                        zip_ref.extractall(extract_dir)
-                elif file_path.endswith(".tar"):
-                    with tarfile.open(file_path, "r") as tar_ref:
-                        dirname = filename.replace(".tar", "")
-                        extract_dir = self._cache_fs.sep.join(
-                            [self._download_folder, record, dirname]
-                        )
-                        tar_ref.extractall(path=extract_dir)
-                elif file_path.endswith(".tar.gz"):
-                    with tarfile.open(file_path, "r:gz") as tar_ref:
-                        dirname = filename.replace(".tar.gz", "")
-                        extract_dir = self._cache_fs.sep(
-                            [self._download_folder, record, dirname]
-                        )
-                        tar_ref.extractall(path=extract_dir)
-                self._events[i].update("Decompressed", np.nan, "")
-                if self._is_cancelled:
-                    break
-
-        self._thread_decompress = threading.Thread(
-            target=decompress, daemon=True, name="decompress_data"
-        )
-        self._thread_decompress.start()
-
-    def _prepare_data(self, *data_ids, **preload_params):
-        def prepare():
-            for i, data_id in enumerate(data_ids):
-                while not self._events[i].status == "Decompressed":
-                    time.sleep(1)
-                self._events[i].update("File processing started", np.nan, "")
-                record, filename = data_id.split("/")
-                format_id = identify_file_format(data_id)
-                filename_unzip = filename.replace(f".{format_id}", "")
-                extract_dir = self._cache_fs.sep.join(
-                    [self._download_folder, record, filename_unzip]
-                )
-                dss = []
-                if is_zarr_directory(extract_dir, self._cache_fs):
-                    dss.append(
-                        self._cache_store.open_data(
-                            extract_dir,
-                            opener_id=f"dataset:zarr:{self._cache_store.protocol}",
-                        )
-                    )
-                    if not filename_unzip.endswith(".zarr"):
-                        filename_unzip = f"{filename_unzip}.zarr"
-                else:
-                    sub_files = self._cache_fs.listdir(extract_dir)
-                    for sub_file in sub_files:
-                        sub_data_id = sub_file["name"].replace(
-                            f"{self._cache_root}/", ""
-                        )
-                        if not self._cache_store.has_data(sub_data_id):
-                            LOG.debug(
-                                f"File with data ID {sub_data_id} cannot be opened, "
-                                f"and thus will not be considered."
-                            )
-                        dss.append(self._cache_store.open_data(sub_data_id))
-                if len(dss) == 1:
-                    self._cache_store.write_data(
-                        dss[0],
-                        f"{record}/{filename_unzip}",
-                        writer_id=f"dataset:zarr:{self._cache_store.protocol}",
-                    )
-                elif preload_params.get("merge"):
-                    ds = xr.merge(dss)
-                    self._cache_store.write_data(
-                        ds,
-                        data_id,
-                        writer_id=f"dataset:zarr:{self._cache_store.protocol}",
-                    )
-                else:
-                    for ds, sub_file in zip(dss, sub_files):
-                        sub_fname = sub_file["name"].split("/")[-1]
-                        data_id = (
-                            f"{record}/{filename_unzip}/"
-                            f"{'.'.join(sub_fname.split('.')[:-1])}.zarr"
-                        )
-                        self._cache_store.write_data(
-                            ds,
+        with requests.get(uri, stream=True) as response:
+            _check_requests_response(response)
+            record, filename = data_id.split("/")
+            record_folder = self._cache_fs.sep.join([self._download_folder, record])
+            if not self._cache_fs.isdir(record_folder):
+                self._cache_fs.makedirs(record_folder)
+            download_path = self._cache_fs.sep.join([record_folder, filename])
+            with open(download_path, "wb") as file:
+                for chunk in response.iter_content(chunk_size=chunk_size):
+                    file.write(chunk)
+                    download_size += len(chunk)
+                    self.notify(
+                        PreloadState(
                             data_id,
-                            writer_id=f"dataset:zarr:{self._cache_store.protocol}",
+                            progress=PRELOAD_DOWNLOAD_FRACTION
+                            * download_size
+                            / total_size,
                         )
-                        LOG.info(
-                            f"Merge is set to False. The sub-dataset is "
-                            f"written to {data_id}"
-                        )
-                self._events[i].update("Preloaded", np.nan, "")
-                if self._is_cancelled:
-                    break
-            self.close()
+                    )
 
-        self._thread_prepare = threading.Thread(
-            target=prepare, daemon=True, name="prepare_data"
+    def _decompress_data(self, data_id):
+        self.notify(
+            PreloadState(
+                data_id,
+                progress=PRELOAD_DOWNLOAD_FRACTION,
+                message="Decompression in progress",
+            )
         )
-        self._thread_prepare.start()
+        record, filename = data_id.split("/")
+        file_path = self._cache_fs.sep.join([self._download_folder, record, filename])
+        if zipfile.is_zipfile(file_path):
+            with zipfile.ZipFile(file_path, "r") as zip_ref:
+                dirname = filename.replace(".zip", "")
+                extract_dir = self._cache_fs.sep.join(
+                    [self._download_folder, record, dirname]
+                )
+                zip_ref.extractall(extract_dir)
+        elif file_path.endswith(".tar") or file_path.endswith(".tar.gz"):
+            format_ext = identify_compressed_file_format(file_path)
+            mode = "r" if format_ext == "tar" else "r:gz"
+            with tarfile.open(file_path, mode) as tar_ref:
+                dirname = filename.replace(f".{format_ext}", "")
+                extract_dir = self._cache_fs.sep.join(
+                    [self._download_folder, record, dirname]
+                )
+                tar_ref.extractall(path=extract_dir, filter="data")
+        self._cache_fs.delete(file_path)
+        self.notify(
+            PreloadState(
+                data_id,
+                progress=PRELOAD_DOWNLOAD_FRACTION + PRELOAD_DECOMPRESSION_FRACTION,
+            )
+        )
+
+    def _prepare_data(self, data_id):
+        self.notify(
+            PreloadState(
+                data_id,
+                progress=PRELOAD_DOWNLOAD_FRACTION + PRELOAD_DECOMPRESSION_FRACTION,
+                message="Write datasets to Zarr in progress",
+            )
+        )
+        format_ext = identify_compressed_file_format(data_id)
+        data_id_mod = data_id.replace(f".{format_ext}", "")
+        extract_dir = self._cache_fs.sep.join([self._download_folder, data_id_mod])
+        # if is_zarr_directory(extract_dir, self._cache_fs):
+        #     ds = self._cache_store.open_data(extract_dir, opener_id=self._io_id)
+        #     if not data_id_mod.endswith(".zarr"):
+        #         data_id_mod = f"{data_id_mod}.zarr"
+        #     self._cache_store.write_data(
+        #         ds, data_id_mod, writer_id=self._io_id, replace=True
+        #     )
+        # else:
+        sub_files = self._cache_fs.listdir(extract_dir)
+        total_size = sum([sub_file["size"] for sub_file in sub_files])
+        size_count = 0
+        for sub_file in sub_files:
+            sub_data_id = sub_file["name"].replace(f"{self._cache_root}/", "")
+            if self._cache_store.has_data(sub_data_id):
+                ds = self._cache_store.open_data(sub_data_id)
+                format_ext = sub_data_id.split(".")[-1]
+                sub_data_id_mod = sub_data_id.replace(f"{DOWNLOAD_FOLDER}/", "")
+                if len(sub_files) == 1:
+                    sub_data_id_mod = "/".join(sub_data_id_mod.split("/")[:-1])
+                sub_data_id_mod = sub_data_id_mod.replace(format_ext, "zarr")
+                self._cache_store.write_data(
+                    ds, sub_data_id_mod, writer_id=self._io_id, replace=True
+                )
+            size_count += sub_file["size"]
+            self.notify(
+                PreloadState(
+                    data_id,
+                    progress=PRELOAD_DOWNLOAD_FRACTION
+                    + PRELOAD_DECOMPRESSION_FRACTION
+                    + (size_count / total_size) * PRELOAD_PROCESSING_FRACTION,
+                )
+            )
+        shutil.rmtree(extract_dir)
+        self.notify(PreloadState(data_id, progress=1.0, message="Preload finished"))
+
+    def _clean_up(self) -> None:
+        for data_id in self._data_ids:
+            zip_filepath = self._cache_fs.sep.join([self._download_folder, data_id])
+            if self._cache_fs.isfile(zip_filepath):
+                self._cache_fs.delete(zip_filepath)
+            format_ext = identify_compressed_file_format(data_id)
+            data_id_mod = data_id.replace(f".{format_ext}", "")
+            extract_dir = self._cache_fs.sep.join([self._download_folder, data_id_mod])
+            if self._cache_fs.isdir(extract_dir):
+                shutil.rmtree(extract_dir)
 
 
-def is_jupyter():
-    import IPython
-
-    return "ZMQInteractiveShell" in IPython.get_ipython().__class__.__name__
+def _check_requests_response(response: requests.Response) -> None:
+    if not response.ok:
+        raise DataStoreError(response.raise_for_status())